requirements.txt
python-3.11
xarray
netcdf4
matplotlib
hvplot
geoviews
scipy
pytest
geopandas
plotly
odc-geo
rioxarray
rasterio